Improve flame graph performance by caching rows.
Before: https://share.firefox.dev/47ydcmp
After: https://share.firefox.dev/40Y48og

Fixes #844.

The old code was saving time by sorting siblings only for the nodes that
were displayed based on the current (preview) range selection. This made
it cheaper to compute the flame graph for a small range, but it meant
that we had to re-do the sort every time the selection changed.

Now we do the sorting once, based on the entire call node table. This is
expensive but it is a one-time cost (and it's still cheaper than if you
were looking at the entire call tree with the old implementation).
Then we cache the "ordered rows" and don't have to sort again on each
range selection change.