Remove the concept of static fields from the marker schema.
Use an optional description field instead.

This causes less trouble with our Flow types. It's also a clearer API.
And it will make it easier to switch marker.data from being an object
to it being an array, in field order, because we'll be able to say that
there should be one array element per field.